requirements.txt
apache-airflow
requests
kafka-python
cassandra-driver 
spark
pyspark